requirements.txt
Django
python-dotenv
Pillow
whitenoise
